userservice/src/main/java/com/clone/userservice/repository/MemoryMemberRepository.java
package com.clone.userservice.repository;

import com.clone.userservice.domain.Member;

import java.util.*;

public class MemoryMemberRepository implements MemberRepository {

    private static Map<Integer, Member> store = new HashMap<>();
    private static int sequence = 0;

    @Override
    public Member save(Member member) {
        member.setID(++sequence);
        store.put(member.getID(), member);
        return member;
    }

    @Override
    public Optional<Member> findByID(int id) {
        return Optional.ofNullable(store.get(id));
    }

    @Override
    public Optional<Member> findByName(String name) {
        return store.values().stream()
                .filter(member -> member.getName().equals(name))
                .findAny();
    }

    @Override
    public List<Member> findAll() {
        return new ArrayList<>(store.values());
    }

    public void clearStore() {
        store.clear();
    }
}
